refactor: consolidate logger utilities into single module
- Move loggerUtils from utils/ to logger/ directory
- Update import paths in affected files
- Simplify logger module structure
